pyproject.toml
[project]
name = "auto-battlebot"
version = "0.1.0"
description = "Auto Battlebot Project"
requires-python = ">=3.10"
dependencies = [
    "ultralytics",
    "numpy>=1.26.0,<2.0.0; platform_machine == \"aarch64\"",
    "numpy>=2.0.0,<3.0.0; platform_machine != \"aarch64\"",
    "matplotlib",
    "scipy",
    "minigencpp @ git+https://github.com/Woz4tetra/minigencpp.git",
    "tqdm>=4.60.0",
    "tomli; python_version < '3.11'",
    "tomli_w",
    "livelossplot",
    "ipython",
    "torchmetrics",
    "segmentation-models-pytorch",
    "opencv-python>=4.5.0",
    "pyyaml>=6.0",
    "onnxscript",
    "natsort",
    "google-auth",
    "google-auth-oauthlib",
    "google-api-python-client",
    "mcap",
    "trimesh",
    # x86 only: often unavailable or different on Jetson (aarch64)
    "torch>=2.0.0; platform_machine == \"x86_64\"",
    "tensorrt==10.14.1.48.post1; platform_machine == \"x86_64\"",
]

[project.optional-dependencies]
floor = [
    "aiofiles>=23.2.1",
    "av>=12.0.0",
    "beautifulsoup4>=4.12.3",
    "lxml>=5.2.2",
    "requests>=2.32.3",
    "scikit-image>=0.24.0",
    "shapely>=2.0.6",
    "tenacity>=9.0.0",
    "xxhash>=3.5.0",
    "yt-dlp>=2025.3.31",
]

[build-system]
requires = ["setuptools>=45", "wheel"]
build-backend = "setuptools.build_meta"

[tool.setuptools]
py-modules = []

[tool.setuptools.packages.find]
where = ["."]
include = []
